+++ Sources/Swift/Helper/SentryEnabledFeaturesBuilder.swift
@@ -49,7 +49,7 @@ import Foundation
         if options.experimental.enableUnhandledCPPExceptionsV2 {
             features.append("unhandledCPPExceptionsV2")
         }
-        if options.enableMetrics {
+        if options.enableMetricsValue {
             features.append("metrics")
         }
         if options.maxFeatureFlags != defaultMaxScopeFeatureFlags {