+++ Tests/SentryObjCTests/SentryObjCLoggerTests.m
@@ -14,9 +14,6 @@
     [SentryObjCSDK startWithConfigureOptions:^(SentryObjCOptions *options) {
         options.dsn = @"https://key@sentry.io/123";
         options.enableCrashHandler = NO;
-#if !SDK_V10
-        options.enableLogs = YES;
-#endif // !SDK_V10
         options.beforeSendLog = ^SentryObjCLog *(SentryObjCLog *log) {
             weakSelf.capturedLog = log;
             return log;
@@ -870,7 +867,10 @@
     [SentryObjCSDK startWithConfigureOptions:^(SentryObjCOptions *options) {
         options.dsn = @"https://key@sentry.io/123";
         options.enableCrashHandler = NO;
+#    pragma clang diagnostic push
+#    pragma clang diagnostic ignored "-Wdeprecated-declarations"
         options.enableLogs = NO;
+#    pragma clang diagnostic pop
     }];
 
     // -- Act & Assert (no crash) --