+++ Tests/SentryTests/Integrations/Metrics/SentryMetricsIntegrationTests.swift
@@ -588,7 +588,7 @@ class SentryMetricsIntegrationTests: XCTestCase {
             $0.dsn = TestConstants.dsnForTestCase(type: Self.self)
             $0.removeAllIntegrations()
 
-            $0.enableMetrics = isEnabled
+            $0.enableMetricsValue = isEnabled
 
             configure?($0)
         }
@@ -599,7 +599,7 @@ class SentryMetricsIntegrationTests: XCTestCase {
         let options = Options()
         options.dsn = TestConstants.dsnForTestCase(type: Self.self)
         options.removeAllIntegrations()
-        options.enableMetrics = true
+        options.enableMetricsValue = true
 
         configure?(options)