+++ Tests/SentryTests/SentrySDKTests.swift
@@ -605,7 +605,7 @@ class SentrySDKTests: XCTestCase {
     
     func testFlush_CallsLoggerCaptureLogs() {
         #if !SDK_V10
-        fixture.client.options.enableLogs = true
+        fixture.client.options.enableLogsValue = true
         #endif // !SDK_V10
         SentrySDKInternal.setCurrentHub(fixture.hub)
         SentrySDK.setStart(with: fixture.client.options)
@@ -626,7 +626,7 @@ class SentrySDKTests: XCTestCase {
     
     func testClose_CallsLoggerCaptureLogs() {
         #if !SDK_V10
-        fixture.client.options.enableLogs = true
+        fixture.client.options.enableLogsValue = true
         #endif // !SDK_V10
         SentrySDKInternal.setCurrentHub(fixture.hub)
         SentrySDK.setStart(with: fixture.client.options)